+++ src/include/sll/types.h
@@ -694,4 +694,11 @@ typedef sll_runtime_object_t* (*sll_quaternary_operator_t)(sll_runtime_object_t*
 
 
 
+typedef struct __SLL_INTEGER_LIST{
+	sll_integer_t* v;
+	sll_array_length_t l;
+} sll_integer_list_t;
+
+
+
 #endif